Added ft_build_branch.c and ft_build_branch_utils.c due to refactoring of functions
