+++ Makefile
@@ -25,7 +25,8 @@ SRCS = $(SRCS_DIR)/main.c \
 		$(SRCS_DIR)/tokenize/ft_process_token_list_utils.c \
 		$(SRCS_DIR)/tokenize/ft_find_next_quote.c \
 		$(SRCS_DIR)/tokenize/ft_build_tree.c \
-		$(SRCS_DIR)/tokenize/ft_build_tree_utils.c \
+		$(SRCS_DIR)/tokenize/ft_build_branch.c \
+		$(SRCS_DIR)/tokenize/ft_build_branch_utils.c \
 		$(SRCS_DIR)/tokenize/ft_free_tree.c \
 		$(SRCS_DIR)/builtins/ft_env.c \
 		$(SRCS_DIR)/builtins/ft_cd.c \
