+++ includes/tokenize.h
@@ -91,11 +91,12 @@ char	*ft_expand_env(char *s, int *i, char **my_envp);
 
 //ft_build_tree.c
 void	*ft_build_tree(t_list **token_list);
+
+//ft_build_branch.c
 void	*ft_build_branch(t_list **list, t_exec *exec);
-//ft_build_tree_utils.c
+//ft_branch_tree_utils.c
 char	**ft_add_to_vector(char **vector, char *new_str);
 char	**ft_get_args(t_list **list);
-bool	ft_find_next_pipe(t_list **list);
 bool	ft_find_next_redir(t_list **list);
 bool	ft_find_next_exec(t_list **list);
